+++ packages/rulesets/src/asyncapi/index.ts
@@ -9,12 +9,15 @@ import {
 } from '@stoplight/spectral-functions';
 
 import asyncApi2ChannelParameters from './functions/asyncApi2ChannelParameters';
+import asyncApi2ChannelServers from './functions/asyncApi2ChannelServers';
 import asyncApi2DocumentSchema from './functions/asyncApi2DocumentSchema';
+import asyncApi2MessageExamplesValidation from './functions/asyncApi2MessageExamplesValidation';
 import asyncApi2OperationIdUniqueness from './functions/asyncApi2OperationIdUniqueness';
 import asyncApi2SchemaValidation from './functions/asyncApi2SchemaValidation';
 import asyncApi2PayloadValidation from './functions/asyncApi2PayloadValidation';
 import asyncApi2ServerVariables from './functions/asyncApi2ServerVariables';
 import { uniquenessTags } from '../shared/functions';
+import asyncApi2Security from './functions/asyncApi2Security';
 
 export default {
   documentationUrl: 'https://meta.stoplight.io/docs/spectral/docs/reference/asyncapi-rules.md',
@@ -68,6 +71,17 @@ export default {
       given: ['$.channels.*', '$.components.channels.*'],
       then: {
         function: asyncApi2ChannelParameters,
+      },
+    },
+    'asyncapi-channel-servers': {
+      description: 'Channel servers must be defined in the "servers" object.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$',
+      then: {
+        function: asyncApi2ChannelServers,
       },
     },
     'asyncapi-headers-schema-type-object': {
@@ -157,6 +171,31 @@ export default {
         function: truthy,
       },
     },
+    'asyncapi-message-examples': {
+      description: 'Examples of message object should follow by "payload" and "headers" schemas.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: [
+        // messages
+        '$.channels.*.[publish,subscribe].message',
+        '$.channels.*.[publish,subscribe].message.oneOf.*',
+        '$.components.channels.*.[publish,subscribe].message',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*',
+        '$.components.messages.*',
+        // message traits
+        '$.channels.*.[publish,subscribe].message.traits.*',
+        '$.channels.*.[publish,subscribe].message.oneOf.*.traits.*',
+        '$.components.channels.*.[publish,subscribe].message.traits.*',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*.traits.*',
+        '$.components.messages.*.traits.*',
+        '$.components.messageTraits.*',
+      ],
+      then: {
+        function: asyncApi2MessageExamplesValidation,
+      },
+    },
     'asyncapi-operation-description': {
       description: 'Operation "description" must be present and non-empty string.',
       recommended: true,
@@ -186,6 +225,20 @@ export default {
       then: {
         field: 'operationId',
         function: truthy,
+      },
+    },
+    'asyncapi-operation-security': {
+      description: 'Operation have to reference a defined security schemes.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$.channels[*][publish,subscribe].security.*',
+      then: {
+        function: asyncApi2Security,
+        functionOptions: {
+          objectType: 'Operation',
+        },
       },
     },
     'asyncapi-parameter-description': {
@@ -351,6 +404,20 @@ export default {
         function: pattern,
         functionOptions: {
           notMatch: 'example\\.com',
+        },
+      },
+    },
+    'asyncapi-server-security': {
+      description: 'Server have to reference a defined security schemes.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$.servers.*.security.*',
+      then: {
+        function: asyncApi2Security,
+        functionOptions: {
+          objectType: 'Server',
         },
       },
     },